requirements.txt
pandas
mysql-connector-python
requests
beautifulsoup4
matplotlib
seaborn
plotly
jupyter